[JENKINS-40617-Handling_unsupported_inputs] Implement same functionality of parameter to input parameter. - Revert failing standard parameter render but returning an alert an any unsupported. - implement a sanity check in parameterRunButton that will not allow to procced but link back to classic build start - move UrlUtils from dashboard to core - extend config aka AppConfig to expose rootUrl and such, so you can now import the config from core and not anymore from local dashboard implementation - add logging to component involved
