+++ blueocean-core-js/src/js/index.js
@@ -100,3 +100,26 @@ export { TimeManager } from './utils/serverBrowserTimeHarmonize';
 export { TimeHarmonizer } from './components/TimeHarmonizer';
 import LiveStatusIndicator from './components/LiveStatusIndicator';
 export { LiveStatusIndicator };
+export {
+    buildOrganizationUrl,
+    buildPipelineUrl,
+    rootPath,
+    buildClassicConfigUrl,
+    buildClassicInputUrl,
+    buildClassicBuildUrl,
+    buildRunDetailsUrl,
+    doubleUriEncode,
+    fetchAllSuffix,
+    applyFetchAll,
+    calculateFetchAll,
+    calculateLogView,
+    calculateLogUrl,
+    calculateNodeBaseUrl,
+    calculateStepsBaseUrl,
+    calculateRunLogURLObject,
+    paginateUrl,
+    endSlash,
+    getRestUrl,
+    buildUrl,
+    relativeUrl,
+} from './utils/UrlUtils';
