+++ blueocean-dashboard/src/main/js/config.js
@@ -1,3 +1,4 @@
+
 //
 // General "system" config information.
 //
