+++ blueocean-dashboard/src/main/less/forms/index.less
@@ -47,6 +47,13 @@
         }
 
         .Dialog-content-margin {
+            a {
+                color: #4A90E2;
+            }
+            a:hover {
+                color: #1f69c1;
+                text-decoration: underline;
+            }
             margin: 24px 32px 16px 32px;
         }
