+++ blueocean-dashboard/src/main/resources/jenkins/plugins/blueocean/dashboard/Messages.properties
@@ -131,3 +131,5 @@ rundetail.tests.results.skipped.count=Skipped - {0}
 rundetail.tests.results.skipped=Skipped
 rundetail.tests.skipped=Skipped - {0}
 Show.complete.logs=Show complete log
+inputStep.error.message=This pipeline uses input types that are unsupported. Use [Jenkins Classic]({0}) to resolve this input step.
+inputStep.error.title=Error
