+++ blueocean-web/src/main/resources/jenkins/plugins/blueocean/web/Messages.properties
@@ -27,6 +27,9 @@ common.state.unstable = Passed in {0} but marked as unstable.
 common.state.failure = Failed in {0}
 common.state.not_built = Not build
 common.state.unknown = Unknow
-parameter.error.message=This pipeline uses input types that are unsupported. {0}
-parameter.error.title=Error
 # the parameterMessage keys assume that you will parse them with a markdown parser
+parameter.error.message=The "{0}" input type is not yet supported.
+parameter.error.title="Error"
+# the parameterMessage keys assume that you will parse them with a markdown parser
+inputParameter.error.message="This pipeline uses input types that are unsupported. Use [Jenkins Classic]({0}) to resolve parametrized build."
+inputParameter.error.title="Error"
